netboot: Attempt booting from all interfaces
The default of booting from eth0 doesn't work that well on devices that
have multiple interfaces, or would boot from non-ethernet interfaces.
This adds a mode where the available interfaces are autodetected and all
tried.

This can make the boot time longer on devices with multiple interfaces
(+30s per interface), so to mitigate it also lowers the interface wait
time. I don't think any reasonable interface would take 30s to boot or
even 10s; in the future we can make this timeout configurable

A way to reduce this wait time would be to start by initializing all
interfaces, then go over the interfaces as soon as they appear up